Renamed Coordinate to Coordinate2D. Coordinate3D will have a separate header/src pair.
