+++ src/morphac/constructs/test/coordinate2D_test.cc
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 
-#include "constructs/include/coordinate.h"
+#include "constructs/include/coordinate2D.h"
 
 namespace {
